[OV-154 #3] Создан новый контроллер для настроек
