+++ orchestrator_service/src/main/java/com/override/orchestrator_service/controller/rest/SettingsController.java
@@ -39,9 +39,4 @@ public class SettingsController {
         backupUserDataService.writingDataFromBackupFile(backupUserDataDTO, telegramUtils.getTelegramId(principal));
         return ResponseEntity.ok(HttpStatus.ACCEPTED);
     }
-
-    @GetMapping("/environment/telegramBotName")
-    public String getTelegramBotName() {
-        return telegramUtils.getTelegramBotName();
-    }
 }
