+++ orchestrator_service/src/main/java/com/override/orchestrator_service/controller/rest/TransactionController.java
@@ -4,7 +4,6 @@ import com.override.dto.TransactionDTO;
 import com.override.dto.TransactionDefineDTO;
 import com.override.dto.TransactionMessageDTO;
 import com.override.dto.TransactionResponseDTO;
-import com.override.orchestrator_service.config.LongPollingOverviewProperties;
 import com.override.orchestrator_service.mapper.TransactionMapper;
 import com.override.orchestrator_service.model.Transaction;
 import com.override.orchestrator_service.service.DefineService;
@@ -40,9 +39,6 @@ public class TransactionController {
 
     @Autowired
     private DefineService defineService;
-
-    @Autowired
-    private LongPollingOverviewProperties longPollingOverviewProperties;
 
     @PostMapping("/transaction")
     public TransactionResponseDTO processTransaction(@RequestBody TransactionMessageDTO transactionMessage, Principal principal) throws InstanceNotFoundException {
@@ -104,10 +100,5 @@ public class TransactionController {
     @DeleteMapping("/transaction/{id}")
     public void deleteTransactionById(@PathVariable("id") UUID id) {
         transactionService.deleteTransactionById(id);
-    }
-
-    @GetMapping("/transaction/longPolling")
-    public Integer getLongPollingData() {
-        return longPollingOverviewProperties.getPeriodOfInactivity();
     }
 }
