+++ orchestrator_service/src/main/resources/static/scripts/history_scripts.js
@@ -20,7 +20,7 @@ $(window).scroll(function () {
 function getTelegramBotName() {
     return new Promise(function (resolve, reject) {
         $.ajax({
-            url: '/settings/environment/telegramBotName',
+            url: '/properties/telegramBotName',
             method: 'GET',
             dataType: 'text',
             success: function (response) {
