+++ orchestrator_service/src/main/resources/static/scripts/overview_scripts.js
@@ -34,7 +34,7 @@ let maxSingleTransactionAmount;
 function getTelegramBotName() {
     return new Promise(function (resolve, reject) {
         $.ajax({
-            url: '/settings/environment/telegramBotName',
+            url: '/properties/telegramBotName',
             method: 'GET',
             dataType: 'text',
             success: function (response) {
@@ -138,7 +138,7 @@ function getPeriodOfInactivity() {
     let periodOfInactivity;
     $.ajax({
         method: 'GET',
-        url: './transaction/longPolling',
+        url: './properties/longPolling',
         contentType: "application/json; charset=utf8",
         async: false,
         success: function (data) {
